+++ sql/starting_items.sql
@@ -19,16 +19,3 @@ CREATE TRIGGER IF NOT EXISTS starting_items_trigger
         WHERE
             profession IS NULL OR profession = NEW.starting_profession;
     END;
-
-INSERT OR IGNORE INTO starting_items (profession, item_template_id, quantity, equipped) VALUES
-(NULL, 5588, NULL, NULL);  -- Tutorial Guide
-
--- For test purposes only:
-INSERT OR IGNORE INTO starting_items (profession, item_template_id, quantity, equipped) VALUES
-(NULL, 6408, NULL, TRUE),  -- Formal Wear
-(NULL, 6621, NULL, FALSE), -- Infinity Spear
-(NULL, 6611, NULL, TRUE),  -- Infinity Blade
-(NULL, 6377, NULL, TRUE),  -- Imperial Crusader Shield
-(NULL, 6618, NULL, FALSE), -- Infinity Fangs
-(NULL, 6619, NULL, FALSE), -- Infinity Bow
-(NULL, 1345, 100,  FALSE); -- Arrow (S-grade)